Object fields are now stored with class_name/field_name as their key in JavaObject.fields, where class_name is the class that 'owns' that field.
This works because we do field lookup before performing a get/set field, so we already know the class that owns the field.

For native methods/creating objects with predefined fields, we need to look up which class owns the field manually (unless you want to suffer a dynamic field lookup at runtime and run a field lookup).