Load exception handler classes upon method lookup.
Since exception handlers need to check if the thrown exception is a
subclass of their handler class, their handler classes must be loaded in
the handler's classloader. We could conceivably do this at the time when
the handler is actually being evaluated against a thrown exception, but
HotSpot seems to do it at method lookup time, and that seemed easier to
implement Doppio.

Note that I did not change all the method_lookup calls to fall back to
resolve_method in case of failure. Given that all the tests pass, I
think we should be okay. However, we really should get a compiler to
generate this ugly code boilerplate for us. Moreover, it would be able
to safely bypass the async fallback if the method has no handlers at
all.

The small snippet in #176 now works correctly, but JRuby 1.7.3 still
fails to load its REPL, exiting quietly instead.