Welp, static isn't exactly what I thought it was.
A static variable is *not* defined on the prototype -- it's defined on the constructor object.

This is... somewhat maddening. Why?

Say we have an interface that defines the property 'foo'. If you have a class that defines 'foo' statically *or* non-statically, it satisfies the interface... but the syntax for accessing 'foo' differs! :(

Anyway, reverting the static 'enhancements' I made; back to defining the same string on a bunch of objects. :(