Initial build of frontend that uses the doppio library.
Release build works! Dev build probably doesn't, but I'll fix that.

The frontend is grabbing a copy of util, and, consequently, gLong. That's because util actually references gLong in one of its methods. I can do the stub trick to remove the gLong dependency. Util will need to be duplicated for now, since the JVM and the browser use the same methods in it.

I have an idea for removing the need for `doppio_stub` and `jvm_stub`... so stay tuned.