Experimental program to generate native method templates, in JavaScript or TypeScript.
Note that these templates are unusable until we migrate the native methods
into a form that can be dynamically loaded at runtime.

I also need to modify code generation to export a variable with a known
name that contains everything defined within the template. Then I need to
test that this feasibly works.