Migrating gruntfile logic to Java 8, updating Javap util  for Java 8.
Note: At least in Windows, the current JDK does *not* add tools.jar to
the classpath, resulting in failed compilations unless you explicitly add
it. Our Gruntfile explicitly informs javac to use our bootclasspath, which
does include tools.jar.

As a result, if you:
* Try to manually compile classes/util/Javap.java w/o tools.jar, or
* Try to run classes.util.Javap w/o tools.jar

...it will fail because it won't be able to find com.sun.tools.Javap.

In good news, Javap is saner to interact with.