Fixing assertion failure dev build of doppio w/ invokeBasic
invokeBasic is typed with an Object return value, but it's a special JVM function that can return whatever it pleases.

Normally, this would be a threat to JVM integrity, but invokeBasic is a private method that is only invoked by trusted JVM code. (Not that that makes me feel any better, but my hands are tied.)