Miscellaneous cleanups and skeleton for the new UnixFileSystem class.
For those following at home, make sure you delete and re-generate includes/, as the old JVMTypes does not include some definitions that the UnixFileSystem natives require.

Committing this as a checkpoint before I fix a bug with our thread scheduler. There's a code path in the JDK triggered by the FileWrite test (only when run in the browser) that is the perfect adversary for our thread scheduler.

The thread scheduler is approximately round robin... but not quite. I'm going to overhaul it so it is actually round robin, and it should simplify its logic. This is long overdue.