Use a JAR-based JDK, and use BrowserFS throughout for handling them.
Significantly cleans up JAR file support, removes dependency on node-zip (and the requirement for node builds to fully extract JAR files before running them), and also fixes up Package-based natives for full Package object support.

Note that the JAR files in this JDK are not compressed, so it should not actually cause a slowdown (in fact, the opposite *should* be true...)...

However... it currently causes a slowdown, as the JDK will derpingly re-open all JAR files on the classpath in Java land whenever system packages are requested. If we replace Jazzlib, this will not be a problem...