Simplify DoppioJVM configuration, test in a WebWorker, fix Safari.
DoppioJVM now only needs you to tell it where doppioHome is.

The Safari bug: Safari doesn't have .slice on TypedArrays, but it has it on ArrayBuffers. I fixed the code to slice on ArrayBuffers.

The test runner (grunt test) runs in a WebWorker if available. Weirdly, WebWorker tests in Safari seem to die after the 15th test... which I cannot figure out, since non-Chrome browsers do not have WebWorker debugging (...and it's 2015! Dear god, why??). I'll have to do some good ole printf debugging to sort that out.

Also, the WebWorker tests appear to run much more slowly in Chrome than in the main thread. This could be due to our setImmediate polyfill.

Finally, our build script is beginning to get unwieldy... so I may switch to something cleaner soon that supports incremental builds out-of-the-box for all of our weird tasks.