Making benchmarks easy to run.
grunt benchmark-node to run benchmarks in node (interpreter-only + JIT) and write results to `node.json` and `node-int.json`,
grunt benchmark-native to run benchmarks in native Java (interpreter-only) and write results to `native_java.json`
grunt benchmark-browser to start a server that runs benchmarks in the main browser context and a webworker, with interpreter-only and JIT. Browser POSTs results to server, server writes results to [browsername]-[version](-int?)(-ww?).json.

To use these runners:

Add a folder vendor/benchmark with all of the data needed for benchmarking. Include a file benchmark.json which specifies the names of benchmarks and the commands to run them.

Example:

{
  "kawa-scheme NQueens": {
    "args": ["-jar", "programs/kawa-2.0.jar", "-f", "fixtures/kawa/nqueens2.sch"]
  },
  "Rhino: binary-trees and controlflow-recursive": {
    "cwd": "fixtures/rhino",
    "args": ["-jar", "../../programs/rhino1.7.6.jar", "-f", "run.js"]
  }
}