Trying to get sockets working again.
Sockets are abstracted in `socket.ts`, and I have implemented them for Node and Websockify. This cleans up our native methods.

The TCPTest works in Node but fails in the browser. The Websockify JavaScript library needs some loving. It assumes that it is running in the main browser context and spews global variables all over the place. I've made some adjustments, but it seems like it may be worth reimplementing in TypeScript in a more modular fashion. I have it working in the Worker context for the unit tests, but it fails to properly do the Websockify handshake.