Additional java.nio file natives, tests, and bug fixes.
The NioFilesPath test fails because it reads an extra line from a file. After much debugging, I discovered that it's a bug in how we manually track file descriptors' positions around the codebase. I'm not sure if that hack is even necessary, so I may try ripping it out and seeing what happens.

Fixes #472.